implify a lot of things by only considering up to Integer.MAX_VALUE worth of meet when creating Concoctions; using more might well result in creating more items than KoL itself is willing to have inventory.
git-svn-id: http://svn.code.sf.net/p/kolmafia/code@19631 b29ace70-8910-0410-8dcc-aa2fc6433167

+++ src/net/sourceforge/kolmafia/maximizer/CheckedItem.java
@@ -304,7 +304,7 @@ public class CheckedItem
 
 	public int inventory;
 	public int initial;
-	public long creatable;
+	public int creatable;
 	public int npcBuyable;
 	public int mallBuyable;
 	public int foldable;
